+++ frontend/src/components/Report.vue
@@ -884,6 +884,7 @@ import CryptoJS from 'crypto-js'
 import QrcodeVue from 'qrcode.vue'
 import csa from './CSA.vue'
 import moment from 'moment'
+import histovec from '../assets/js/histovec'
 
 export default {
   components: {
@@ -1058,198 +1059,6 @@ export default {
       n = n + ''
       return n.length >= width ? n : new Array(width - n.length + 1).join(z) + n
     },
-    formatDate (isoDate) {
-      let date = new Date(isoDate)
-      let mm = date.getMonth() + 1 // getMonth() is zero-based
-      let dd = date.getDate()
-
-      return [(dd > 9 ? '' : '0') + dd,
-        (mm > 9 ? '' : '0') + mm,
-        date.getFullYear()
-      ].join('/')
-    },
-    histoFilter (historique) {
-      let h = historique.filter(event => this.operations[event.opa_type] !== undefined)
-      h = this.$lodash.orderBy(h, ['opa_date'], ['desc'])
-      return h.map(event => {
-        return {'date': this.formatDate(event.opa_date), 'nature': this.operations[event.opa_type]}
-      })
-    },
-    calcNbTit (historique) {
-      let opTit = ['IMMAT_NORMALE', 'IMMAT_NORMALE_PREM_VO', 'CHANG_LOC', 'CHANG_LOC_CVN', 'CHANG_TIT_NORMAL', 'CHANG_TIT_NORMAL_CVN']
-      let nbTit = historique.filter(event => opTit.includes(event.opa_type))
-      return nbTit.length
-    },
-    monthDiff (d1, d2) {
-      var months
-      months = (d2.getFullYear() - d1.getFullYear()) * 12
-      months -= d1.getMonth() + 1
-      months += d2.getMonth()
-      // increment months if d2 comes later in its month than d1 in its month
-      if (d2.getDate() >= d1.getDate()) { months++ }
-      return months <= 0 ? 0 : months
-    },
-    calcCertifDepuis (dateStr) {
-      // Si on détecte que la date est au format FR alors on l'a converti
-      if (moment(dateStr, 'DD/MM/YYYY', true).isValid()) {
-        dateStr = moment(dateStr, 'DD/MM/YYYY').format('YYYY-MM-DD')
-      }
-      let nbMonth = this.monthDiff(new Date(dateStr), new Date())
-      if (nbMonth <= 18) {
-        return nbMonth + ' mois'
-      } else {
-        let year = Math.floor(nbMonth / 12)
-        let month = nbMonth - 12 * year
-        if ((month > 0) && (year < 10)) {
-          return (year > 1) ? year + ' ans et ' + month + ' mois' : year + ' an et ' + month + ' mois'
-        } else {
-          return (year > 1) ? year + ' ans' : year + ' an'
-        }
-      }
-    },
-    getVehiculeTypeCarburant (carburant) {
-      // Mapping Carburant
-      let essence = ['ES', 'EH', 'ET', 'FE', 'FH']
-      let diesel = ['GO', 'GA', 'GE', 'GF', 'GG', 'GH', 'PL', 'GQ']
-      let electHydro = ['AC', 'EL', 'H2', 'HE', 'HH']
-      let gaz = ['EG', 'EN', 'EP', 'EQ', 'FG', 'FN', 'G2', 'GN', 'GP', 'GZ', 'NH', 'PH']
-      let hybrideRech = ['EE', 'EM', 'ER', 'FL', 'GL', 'GM', 'NE', 'PE']
-      let typeCarburant = ''
-      if (essence.includes(carburant)) {
-        typeCarburant = 'essence'
-      } else if (diesel.includes(carburant)) {
-        typeCarburant = 'diesel'
-      } else if (electHydro.includes(carburant)) {
-        typeCarburant = 'electrique'
-      } else if (gaz.includes(carburant)) {
-        typeCarburant = 'gaz'
-      } else if (hybrideRech.includes(carburant)) {
-        typeCarburant = 'hybride'
-      }
-      return typeCarburant
-    },
-    getVehiculeLogo (genre) {
-      let moto = ['MTL', 'MTT1', 'MTT2', 'MTTE', 'CL']
-      let truck = ['CAM', 'Deriv-VP', 'TRA', 'TRR', 'TCP']
-      let type = 'car'
-      if (moto.includes(genre)) {
-        type = 'motorcycle'
-      } else if (truck.includes(genre)) {
-        type = 'truck'
-      }
-      return type
-    },
-    getVignetteNumero (genre, categorie, typeCarburant, pollution, datePremImmat) {
-      let splitDate = datePremImmat.split('/')
-      let dateImmatEn = new Date(splitDate[2] + '-' + splitDate[1] + '-' + splitDate[0])
-      let vignette = ''
-      // Mapping Norme Euro
-      let normeEuro = (pollution) ? pollution.split('EURO') : ''
-      let numeroEuro = (normeEuro !== '' && normeEuro[1] !== undefined) ? normeEuro[1] : ''
-      let voitureParticuliere = []
-      let vehiculeUtilitaireLegers = []
-      let motocycle = []
-      let cyclomoteur = []
-      let poidsLourdsAutobusAutocar = []
-      if (typeCarburant === 'gaz' || typeCarburant === 'hybride') {
-        vignette = 1
-      } else if (typeCarburant === 'electrique') {
-        vignette = 'electrique'
-      } else {
-        // Mapping Categorie
-        if ((categorie !== '' && categorie !== undefined)) {
-          let categ = categorie.split('-')
-          categorie = categ[0] // Cas des categories qui contiennent des sous catégories (ex: L3e-A1) on récupère uniquement la première categorie
-          voitureParticuliere = ['M1']
-          vehiculeUtilitaireLegers = ['N1']
-          motocycle = ['L3e', 'L4e', 'L5e', 'L7e']
-          cyclomoteur = ['L1e', 'L2e', 'L6e']
-          poidsLourdsAutobusAutocar = ['M2', 'M3', 'N2', 'N3']
-        } else if (genre !== '') {
-          categorie = genre // Uniquement si la catégorie n'est pas remplie on remplace par genre
-          voitureParticuliere = ['VP']
-          vehiculeUtilitaireLegers = ['CTTE']
-          motocycle = ['QM', 'TM', 'MTL', 'MTT1', 'MTT2', 'MTTE']
-          cyclomoteur = ['CYCL', 'CL']
-          poidsLourdsAutobusAutocar = ['CAM', 'TCP']
-        } else {
-          return vignette
-        }
-
-        if (motocycle.includes(categorie) || cyclomoteur.includes(categorie)) {
-          if (numeroEuro === '4' || (numeroEuro === '' && motocycle.includes(categorie) && dateImmatEn >= new Date('2017-01-01')) || (numeroEuro === '' && cyclomoteur.includes(categorie) && dateImmatEn >= new Date('2018-01-01'))) {
-            vignette = 1
-          } else if (numeroEuro === '3' || (numeroEuro === '' && motocycle.includes(categorie) && (dateImmatEn >= new Date('2007-01-01') && dateImmatEn <= new Date('2016-12-31'))) || (numeroEuro === '' && cyclomoteur.includes(categorie) && (dateImmatEn >= new Date('2007-01-01') && dateImmatEn <= new Date('2017-12-31')))) {
-            vignette = 2
-          } else if (numeroEuro === '2' || (numeroEuro === '' && dateImmatEn >= new Date('2004-07-01') && dateImmatEn <= new Date('2006-12-31'))) {
-            vignette = 3
-          } else if (dateImmatEn >= new Date('2000-06-01') && dateImmatEn <= new Date('2004-06-30')) {
-            vignette = 4
-          }
-        } else if (voitureParticuliere.includes(categorie)) {
-          if (typeCarburant === 'essence') {
-            if (numeroEuro === '5' || numeroEuro === '6' || (numeroEuro === '' && dateImmatEn >= new Date('2011-01-01'))) {
-              vignette = 1
-            } else if (numeroEuro === '4' || (numeroEuro === '' && dateImmatEn >= new Date('2006-01-01') && dateImmatEn <= new Date('2010-12-31'))) {
-              vignette = 2
-            } else if (numeroEuro === '2' || numeroEuro === '3' || (numeroEuro === '' && dateImmatEn >= new Date('1997-01-01') && dateImmatEn <= new Date('2005-12-31'))) {
-              vignette = 3
-            }
-          } else if (typeCarburant === 'diesel') {
-            if (numeroEuro === '5' || numeroEuro === '6' || (numeroEuro === '' && dateImmatEn >= new Date('2011-01-01'))) {
-              vignette = 2
-            } else if (numeroEuro === '4' || (numeroEuro === '' && dateImmatEn >= new Date('2006-01-01') && dateImmatEn <= new Date('2010-12-31'))) {
-              vignette = 3
-            } else if (numeroEuro === '3' || (numeroEuro === '' && dateImmatEn >= new Date('2001-01-01') && dateImmatEn <= new Date('2005-12-31'))) {
-              vignette = 4
-            } else if (numeroEuro === '2' || (numeroEuro === '' && dateImmatEn >= new Date('1997-01-01') && dateImmatEn <= new Date('2000-12-31'))) {
-              vignette = 5
-            }
-          }
-        } else if (vehiculeUtilitaireLegers.includes(categorie)) {
-          if (typeCarburant === 'essence') {
-            if (numeroEuro === '5' || numeroEuro === '6' || (numeroEuro === '' && dateImmatEn >= new Date('2011-01-01'))) {
-              vignette = 1
-            } else if (numeroEuro === '4' || (numeroEuro === '' && dateImmatEn >= new Date('2006-01-01') && dateImmatEn <= new Date('2010-12-31'))) {
-              vignette = 2
-            } else if (numeroEuro === '2' || numeroEuro === '3' || (numeroEuro === '' && dateImmatEn >= new Date('1997-01-01') && dateImmatEn <= new Date('2005-12-31'))) {
-              vignette = 3
-            }
-          } else if (typeCarburant === 'diesel') {
-            if (numeroEuro === '5' || numeroEuro === '6' || (numeroEuro === '' && dateImmatEn >= new Date('2011-01-01'))) {
-              vignette = 2
-            } else if (numeroEuro === '4' || (numeroEuro === '' && dateImmatEn >= new Date('2006-01-01') && dateImmatEn <= new Date('2010-12-31'))) {
-              vignette = 3
-            } else if (numeroEuro === '3' || (numeroEuro === '' && dateImmatEn >= new Date('2001-01-01') && dateImmatEn <= new Date('2005-12-31'))) {
-              vignette = 4
-            } else if (numeroEuro === '2' || (numeroEuro === '' && dateImmatEn >= new Date('1997-01-01') && dateImmatEn <= new Date('2000-12-31'))) {
-              vignette = 5
-            }
-          }
-        } else if (poidsLourdsAutobusAutocar.includes(categorie)) {
-          if (typeCarburant === 'essence') {
-            if (numeroEuro === '6' || (numeroEuro === '' && dateImmatEn >= new Date('2014-01-01'))) {
-              vignette = 1
-            } else if (numeroEuro === '5' || (numeroEuro === '' && dateImmatEn >= new Date('2009-10-01') && dateImmatEn <= new Date('2013-12-31'))) {
-              vignette = 2
-            } else if (numeroEuro === '3' || numeroEuro === '4' || (numeroEuro === '' && dateImmatEn >= new Date('2001-10-01') && dateImmatEn <= new Date('2009-09-30'))) {
-              vignette = 3
-            }
-          } else if (typeCarburant === 'diesel') {
-            if (numeroEuro === '6' || (numeroEuro === '' && dateImmatEn >= new Date('2014-01-01'))) {
-              vignette = 2
-            } else if (numeroEuro === '5' || (numeroEuro === '' && dateImmatEn >= new Date('2009-10-01') && dateImmatEn <= new Date('2013-12-31'))) {
-              vignette = 3
-            } else if (numeroEuro === '4' || (numeroEuro === '' && dateImmatEn >= new Date('2006-10-01') && dateImmatEn <= new Date('2009-09-30'))) {
-              vignette = 4
-            } else if (numeroEuro === '3' || (numeroEuro === '' && dateImmatEn >= new Date('2001-10-01') && dateImmatEn <= new Date('2006-09-30'))) {
-              vignette = 5
-            }
-          }
-        }
-      }
-      return vignette
-    },
     send (e) {
       this.status = 'posting'
       if (this.note || this.notShow) {
@@ -1394,105 +1203,8 @@ export default {
               this.result = 'error'
             }
             try {
-              // filtre l'historique des opérations annulées
-              veh.historique = (veh.historique === undefined) ? [] : veh.historique.filter(event => event.ope_date_annul === undefined)
-              // réordonne l'historique des opérations
-              veh.historique = this.$lodash.orderBy(veh.historique, ['opa_date'])
-              this.v.date_update = veh.date_update || this.v.date_update
-              this.vin = veh.vin
-              this.v.ctec.vin = veh.vin
-              this.plaque = veh.plaq_immat
-              this.v.plaque = veh.plaq_immat
-              this.v.ctec.couleur = veh.couleur || this.default
-              this.v.ctec.cnit = veh.num_cnit
-              this.v.ctec.tvv = veh.tvv
-              this.v.ctec.reception.type = veh.type_reception
-              this.v.ctec.reception.numero = veh.cveh_num_reception
-              this.v.ctec.puissance.cylindres = veh.CTEC_CYLINDREE
-              this.v.ctec.puissance.nette = veh.CTEC_PUISS_NETTE
-              this.v.ctec.puissance.cv = veh.CTEC_PUISS_CV
-              this.v.ctec.puissance.norm = veh.CTEC_RAPPORT_PUIS_MASSE
-              this.v.ctec.places.assis = veh.CTEC_PLACES_ASSISES
-              this.v.ctec.places.debout = veh.CTEC_PLACES_DEBOUT
-              this.v.ctec.db = veh.CTEC_NIVEAU_SONORE
-              this.v.ctec.co2 = veh.CTEC_CO2
-              this.v.ctec.moteur = veh.CTEC_VITESSE_MOTEUR
-              this.v.ctec.marque = veh.marque
-              this.v.ctec.modele = veh.nom_commercial
-              this.v.ctec.genre = veh.CTEC_RLIB_GENRE
-              this.v.ctec.categorie = veh.CTEC_RLIB_CATEGORIE
-              this.v.ctec.carrosserie.national = veh.CTEC_RLIB_CARROSSERIE_NAT
-              this.v.ctec.carrosserie.ce = veh.CTEC_RLIB_CARROSSERIE_CE
-              this.v.ctec.environnement = veh.CTEC_RLIB_POLLUTION
-              this.v.ctec.energie = veh.CTEC_RLIB_ENERGIE
-              this.v.ctec.PT.admissible = veh.pt_tech_adm_f1
-              this.v.ctec.PT.AC = veh.ptac_f2
-              this.v.ctec.PT.RA = veh.ptra_f3
-              this.v.ctec.PT.service = veh.pt_service_g
-              this.v.ctec.PT.AV = veh.ptav_g1
-              this.v.titulaire.identite = [veh.pers_raison_soc_tit, veh.pers_siren_tit, veh.pers_nom_naissance_tit, veh.pers_prenom_tit].join(' ')
-              this.v.titulaire.adresse = (veh.adr_code_postal_tit !== undefined) ? this.pad(veh.adr_code_postal_tit, 5) : this.default
-              this.v.certificat.premier = veh.date_premiere_immat || this.default
-              // véhicule importé: changement de règle de gestion #406
-              this.v.certificat.etranger = (veh.import === 'OUI')
-              this.v.certificat.siv = veh.date_premiere_immat_siv || this.default
-              this.v.certificat.fr = (this.v.certificat.etranger && (veh.historique !== undefined)) ? this.formatDate(veh.historique[0].opa_date) : this.v.certificat.premier
-              this.v.fni = ((veh.dos_date_conversion_siv !== undefined) && (veh.historique !== undefined)) ? ((veh.historique[0].opa_type === 'IMMAT_NORMALE') ? 'converti' : 'converti_incertain') : (veh.date_premiere_immat_siv === undefined)
-              this.v.certificat.incertain = !this.v.certificat.etranger && (this.v.certificat.siv !== this.v.certificat.fr) && (veh.historique[0].opa_type !== 'IMMAT_NORMALE')
-              this.v.certificat.courant = veh.date_emission_CI || this.default
-              // this.v.certificat.depuis = (this.calcCertifDepuis(this.$lodash.orderBy(veh.historique.filter(e => (e.opa_type === 'IMMAT_NORMALE' || e.opa_type === 'IMMAT_NORMALE_PREM_VO' || e.opa_type === 'CHANG_TIT_NORMAL' || e.opa_type === 'CHANG_TIT_NORMAL_CVN')), ['opa_date'], ['desc'])[0].opa_date) || this.calcCertifDepuis(veh.date_premiere_immat))
-              this.v.certificat.depuis = this.calcCertifDepuis((this.$lodash.orderBy(veh.historique.filter(e => (e.opa_type === 'IMMAT_NORMALE' || e.opa_type === 'IMMAT_NORMALE_PREM_VO' || e.opa_type === 'CHANG_TIT_NORMAL' || e.opa_type === 'CHANG_TIT_NORMAL_CVN')), ['opa_date'], ['desc'])[0] || {'opa_date': veh.date_premiere_immat}).opa_date)
-
-              if ((this.v.fni !== true) && (this.v.certificat.fr !== this.v.certificat.siv) && ((veh.historique === undefined) || (!veh.historique.some(e => e.opa_type.match(/(CONVERSION_DOSSIER_FNI|.*_CVN)/))))) {
-                let tmp = veh.historique
-                tmp.push({opa_date: this.v.certificat.siv.replace(/^(..)\/(..)\/(....)$/, '$3-$2-$1'), opa_type: 'CONVERSION_DOSSIER_FNI'})
-                this.v.historique = (veh.historique !== undefined) ? this.histoFilter(tmp) : []
-              } else {
-                this.v.historique = (veh.historique !== undefined) ? this.histoFilter(veh.historique) : []
-              }
-              this.v.nb_proprietaires = veh.nb_proprietaire
-              this.v.nb_tit = (veh.historique !== undefined) ? (this.calcNbTit(veh.historique) + (this.v.certificat.incertain ? 1 : 0)) : undefined
-              this.v.age_veh = veh.age_annee
-              this.v.logo_vehicule = this.getVehiculeLogo(veh.CTEC_RLIB_GENRE)
-              this.v.vignette_numero = this.getVignetteNumero(veh.CTEC_RLIB_GENRE, veh.CTEC_RLIB_CATEGORIE, this.getVehiculeTypeCarburant(veh.CTEC_RLIB_ENERGIE), veh.CTEC_RLIB_POLLUTION, veh.date_premiere_immat)
-
-              this.v.administratif.gages = veh.gage || this.default
-              this.v.administratif.suspension = (veh.suspension === 'NON') ? 'Non' : 'Oui'
-              this.v.administratif.annulation = (veh.annulation_ci === 'NON') ? 'Non' : 'Oui'
-              this.v.administratif.suspensions = (veh.suspension === 'NON') ? ((veh.suspension === 'NON') ? 'NON' : 'certificat annulé') : ((veh.annulation_ci === 'NON') ? 'certificat suspendu' : 'certificat suspendu et annulé') // mapping à valider
-              // opposition et procédure à valider
-              this.v.administratif.otci = (veh.otci === 'NON') ? 'Aucune' : ((veh.otci_pv === 'OUI') ? 'opposition temporaire (PV en attente)' : 'opposition temporaire')
-              this.v.administratif.ove = (veh.ove === 'NON') ? 'Aucune' : 'Oui'
-              this.v.administratif.oppositions = (veh.ove === 'NON') ? ((veh.otci === 'NON') ? 'NON' : (veh.otci_pv === 'OUI') ? 'Opposition temporaire (PV en attente)' : 'opposition temporaire') : ((veh.otci === 'NON') ? 'procédure de réparation contrôlée' : 'opposition temporaire, véhicule endommagé') // mapping à valider
-              this.v.administratif.pv = (veh.otci_pv === 'OUI')
-              // pour l'instant aucun véhicule saisi dans les échantillons
-              this.v.administratif.saisie = (veh.saisie === 'NON') ? 'Aucune' : 'Oui'
-              this.v.administratif.gage = (veh.gage === 'NON') ? 'Aucun' : 'Oui'
-              this.v.administratif.procedures = (veh.saisie === 'NON') ? ((veh.gage === 'NON') ? 'NON' : 'véhicule gagé') : ((veh.annulation_ci === 'NON') ? 'véhicule saisi' : 'véhicule gagé et saisi') // mapping à valider
-              this.v.administratif.vol = veh.vehicule_vole || this.default
-
-              // vol : les informations viennent-elles de foves ?
-              this.v.administratif.titre.vol = veh.ci_vole || this.default
-              this.v.administratif.titre.perte = veh.perte_ci || this.default
-              this.v.administratif.titre.duplicata = (veh.perte_ci === 'OUI') ? 'OUI' : veh.duplicata
-
-              this.v.administratif.synthese = [ 'saisie', 'vehicule_vole', 'gage', 'suspension', 'perte_ci', 'ci_vole', 'annulation_ci', 'duplicata' ].filter(e => (e !== 'duplicata') ? veh[e] === 'OUI' : ((veh['perte_ci'] === 'OUI') || (veh['ci_vole'] === 'OUI') ? false : veh[e] === 'OUI'))
-              if (veh['otci'] === 'OUI') {
-                this.v.administratif.synthese.push(veh['ove'] === 'OUI' ? 'otci_ove' : 'otci')
-              }
-              // véhicule importé : changement de règle de gestion #406
-              this.v.etranger = (veh.import === 'NON') ? 'NON' : [veh.import, veh.imp_imp_immat, veh.pays_import]
-              // ci-dessous : interprétation à confirmer
-              this.v.sinistres = (veh.historique !== undefined) ? (this.$lodash.orderBy(veh.historique.filter(e => (e.opa_type === 'INSCRIRE_OVE') || (e.opa_type === 'DEC_VE')), ['opa_date'], ['desc']).map(e => e.opa_date.replace(/-.*/, ''))) : []
-              this.v.sinistres_nb = (veh.historique !== undefined) ? (this.$lodash.orderBy(veh.historique.filter(e => (e.opa_type === 'INSCRIRE_OVE') || (e.opa_type === 'DEC_VE')), ['opa_date'], ['desc']).map(e => ((e.opa_type === 'INSCRIRE_OVE') ? 10 : 1))) : []
-              this.v.sinistres_nb = this.v.sinistres_nb.length === 0 ? 0 : this.v.sinistres_nb.reduce((a, b) => a + b)
-              this.v.sinistres_nb = Math.max(this.v.sinistres_nb % 10, ((this.v.sinistres_nb - (this.v.sinistres_nb % 10)) / 10))
-              console.log(this.v.sinistres_nb)
-              this.v.sinistre = this.v.sinistres[0]
-              this.v.aptes = (veh.historique !== undefined) ? (this.$lodash.orderBy(veh.historique.filter(e => (e.opa_type === 'LEVER_OVE') || (e.opa_type === 'SEC_RAP_VE')), ['opa_date'], ['desc']).map(e => e.opa_date.replace(/-.*/, ''))) : []
-              this.v.apte = (veh.historique !== undefined) ? ((this.v.aptes[0] > this.v.sinistres[0]) || ((veh.suspension === 'NON') && (veh.ove === 'NON'))) : undefined
+              this.v = histovec.histovec(veh)
               this.result = 'ok'
-              console.log(this.v)
               this.$store.commit('updateSuccess', this.$store.state.success + 1)
               this.$store.commit('updateV', this.v)
               this.$store.commit('updateCode', this.$route.params.code)
